Finish sort, ascend, descend, merge
Divide-and-conquer algorithm for the sorting part, this is a easy way
to implement, add merge_two_sorted function to merge two sorted list
into one, and this function is also useful for the merge function.

Delete all the elements that is less than current one is ascend,
is simple to do in this way, descend is in going from the other
deriction.

Use the queue_context_t mentioned in queue.h for entering each entry
to the queue chain, reuse the merge_two_sorted funtion just in case.